ASR-192: separated create from campaign as its own subcommand
